- moved the backwards compatibility code into a new file
